# openapi.yaml
openapi: 3.1.0
info:
  title: Incident Coordination Engine API
  version: 1.0.0
  description: |
    API v1 pour le client mobile P2 (incident-responder).
    Auth JWT Bearer. Mutations incident via lock optimiste (`version`).

servers:
  - url: /api/v1
    description: API v1 (relative au host ICE)

tags:
  - name: auth
  - name: incidents

paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Obtenir un JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Token émis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/me:
    get:
      tags: [auth]
      security:
        - bearerAuth: []
      summary: Profil courant
      responses:
        "200":
          description: Utilisateur courant
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [user]
                    properties:
                      user:
                        $ref: "#/components/schemas/SessionUser"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /incidents:
    get:
      tags: [incidents]
      security:
        - bearerAuth: []
      summary: Liste paginée (RBAC)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Liste incidents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /incidents/{id}:
    get:
      tags: [incidents]
      security:
        - bearerAuth: []
      summary: Détail incident
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      responses:
        "200":
          description: Incident
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [incident]
                    properties:
                      incident:
                        $ref: "#/components/schemas/Incident"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"

    patch:
      tags: [incidents]
      security:
        - bearerAuth: []
      summary: Mutation incident (lock optimiste)
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchIncidentRequest"
      responses:
        "200":
          description: Incident mis à jour
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [incident]
                    properties:
                      incident:
                        $ref: "#/components/schemas/Incident"
        "409":
          description: Conflit de version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionConflictResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /incidents/{id}/events:
    get:
      tags: [incidents]
      security:
        - bearerAuth: []
      summary: Timeline events
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      responses:
        "200":
          description: Events ordonnés
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [events]
                    properties:
                      events:
                        type: array
                        items:
                          $ref: "#/components/schemas/IncidentEvent"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    IncidentId:
      name: id
      in: path
      required: true
      schema:
        type: string

  responses:
    Unauthorized:
      description: Non authentifié
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Accès refusé
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unprocessable:
      description: Règle métier violée
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    SessionUser:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [responder, lead]

    LoginResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [token, user]
          properties:
            token:
              type: string
            user:
              $ref: "#/components/schemas/SessionUser"

    Incident:
      type: object
      required: [id, title, severity, status, version]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, acknowledged, investigating, resolved]
        version:
          type: integer
        assigneeId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    IncidentEvent:
      type: object
      required: [id, incidentId, type, timestamp]
      properties:
        id:
          type: string
        incidentId:
          type: string
        type:
          type: string
        actorId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object

    PatchIncidentRequest:
      type: object
      required: [version]
      properties:
        version:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [open, acknowledged, investigating, resolved]
        severity:
          type: string
          enum: [low, medium, high, critical]
        assigneeId:
          type: string
          nullable: true
      description: Au moins un parmi status, severity, assigneeId requis.

    IncidentListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Incident"
        meta:
          type: object
          required: [page, pageSize, total, totalPages]
          properties:
            page:
              type: integer
            pageSize:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer

    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"

    VersionConflictResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
        data:
          type: object
          properties:
            incident:
              $ref: "#/components/schemas/Incident"